Update README.md
The purpose of this project is to analyze hotel reservation data. This project will analyze why customers cancel their hotel reservations as part of the hotel's data science team. We will do this by analyzing 20 attributes on 11122 hotel bookings. The report shows the results of our data analysis and provides suggestions for the hotel on how to reduce cancellations from customers. We used three models to achieve this goal: Support Vector Machine, Associative Mining and Partition or Classification